+++ pyvo/registry/tests/test_regtap.py
@@ -8,6 +8,7 @@ from urllib.parse import parse_qsl
 import pytest
 
 from pyvo.registry import search as regsearch
+from pyvo.dal import query as dalq
 
 from astropy.utils.data import get_pkg_data_contents
 
@@ -139,6 +140,23 @@ def datamodel_fixture(mocker):
         yield matcher
 
 
+@pytest.fixture()
+def aux_fixture(mocker):
+    def auxtest_callback(request, context):
+        data = dict(parse_qsl(request.body))
+        query = data['QUERY']
+
+        assert "ivo://ivoa.net/std/tap#aux" in query
+
+        return get_pkg_data_contents('data/regtap.xml')
+
+    with mocker.register_uri(
+        'POST', 'http://dc.g-vo.org/tap/sync',
+        content=auxtest_callback
+    ) as matcher:
+        yield matcher
+
+
 @pytest.mark.usefixtures('keywords_fixture', 'capabilities')
 def test_keywords():
     regsearch(keywords=['vizier', 'pulsar'])
@@ -152,7 +170,7 @@ def test_single_keyword():
 
 @pytest.mark.usefixtures('servicetype_fixture', 'capabilities')
 def test_servicetype():
-    regsearch(servicetype='tap')
+    regsearch(servicetype='table')
 
 
 @pytest.mark.usefixtures('waveband_fixture', 'capabilities')
@@ -163,3 +181,19 @@ def test_waveband():
 @pytest.mark.usefixtures('datamodel_fixture', 'capabilities')
 def test_datamodel():
     regsearch(datamodel='tap')
+
+
+@pytest.mark.usefixtures('aux_fixture', 'capabilities')
+def test_servicetype_aux():
+    regsearch(servicetype='table', includeaux=True)
+
+
+@pytest.mark.usefixtures('aux_fixture', 'capabilities')
+def test_keyword_aux():
+    regsearch(keywords=['pulsar'], includeaux=True)
+
+
+@pytest.mark.usefixtures('aux_fixture', 'capabilities')
+def test_bad_servicetype_aux():
+    with pytest.raises(dalq.DALQueryError):
+        regsearch(servicetype='bad_servicetype', includeaux=True)