+++ CHANGES.rst
@@ -3,6 +3,8 @@
 
 - Make .bysemantics expand its terms to the entire branch by default [#241]
 
+- Added optional includeaux flag for regTAP search() [#258]
+
 
 1.1 (2020-06-26)
 ================